lib/cli: bring back cli_util start_repo_transaction
