+++ cli/src/cli_util.rs
@@ -130,7 +130,6 @@ use jj_lib::str_util::StringExpression;
 use jj_lib::str_util::StringMatcher;
 use jj_lib::transaction::Transaction;
 use jj_lib::transaction::TransactionCommitError;
-use jj_lib::transaction::start_repo_transaction;
 use jj_lib::working_copy;
 use jj_lib::working_copy::CheckoutStats;
 use jj_lib::working_copy::LockedWorkingCopy;
@@ -732,17 +731,13 @@ impl CommandHelper {
                     )?;
                     // TODO: It may be helpful to print each operation we're merging here
                     let transaction_description = "reconcile divergent operations";
-                    let transaction_attributes = [(
-                        "args".to_string(),
-                        command_args_to_transaction_attribute(&self.data.string_args),
-                    )];
                     merge_operations(
                         Some(ui),
                         repo_loader,
                         op_heads,
                         Some(workspace_name),
                         Some(transaction_description),
-                        transaction_attributes,
+                        &self.data.string_args,
                     )
                     .await
                 },
@@ -776,8 +771,9 @@ pub async fn merge_operations(
     operations: Vec<Operation>,
     workspace_name: Option<&WorkspaceName>,
     transaction_description: Option<&str>,
-    transaction_attributes: impl IntoIterator<Item = (String, String)>,
+    command_args: &[String],
 ) -> Result<Operation, CommandError> {
+    let transaction_attributes = command_args_to_transaction_attribute(command_args);
     let (merged_repo, num_rebased) = repo_loader
         .merge_operations(
             operations,
@@ -2061,14 +2057,10 @@ to the current parents may contain changes from multiple commits.
                 .map_err(snapshot_command_error)?
         };
         if new_tree.tree_ids_and_labels() != wc_commit.tree().tree_ids_and_labels() {
-            let transaction_attributes = [(
-                "args".to_string(),
-                command_args_to_transaction_attribute(self.env.command.string_args()),
-            )];
             let mut tx = start_repo_transaction(
                 &self.user_repo.repo,
-                Some(&workspace_name),
-                transaction_attributes,
+                &workspace_name,
+                self.env.command.string_args(),
             );
             tx.set_is_snapshot(true);
             let immutable_expr = self
@@ -2242,14 +2234,10 @@ to the current parents may contain changes from multiple commits.
     }
 
     pub fn start_transaction(&mut self) -> WorkspaceCommandTransaction<'_> {
-        let transaction_attributes = [(
-            "args".to_string(),
-            command_args_to_transaction_attribute(self.env.command.string_args()),
-        )];
         let tx = start_repo_transaction(
             self.repo(),
-            Some(self.workspace_name()),
-            transaction_attributes,
+            self.workspace_name(),
+            self.env.command.string_args(),
         );
         let id_prefix_context = mem::take(&mut self.user_repo.id_prefix_context);
         WorkspaceCommandTransaction {
@@ -2802,7 +2790,23 @@ jj git init",
     }
 }
 
-pub fn command_args_to_transaction_attribute(command_args: &[String]) -> String {
+pub fn start_repo_transaction(
+    repo: &Arc<ReadonlyRepo>,
+    workspace_name: &WorkspaceName,
+    string_args: &[String],
+) -> Transaction {
+    let mut tx = repo.start_transaction();
+    tx.set_workspace_name(workspace_name);
+    for (key, value) in command_args_to_transaction_attribute(string_args) {
+        tx.set_attribute(key, value);
+    }
+    tx
+}
+
+fn command_args_to_transaction_attribute(command_args: &[String]) -> Vec<(String, String)> {
+    if command_args.is_empty() {
+        return vec![];
+    }
     // TODO: Either do better shell-escaping here or store the values in some list
     // type (which we currently don't have).
     let shell_escape = |arg: &String| {
@@ -2827,7 +2831,7 @@ pub fn command_args_to_transaction_attribute(command_args: &[String]) -> String 
     };
     let mut quoted_strings = vec!["jj".to_string()];
     quoted_strings.extend(command_args.iter().skip(1).map(shell_escape));
-    quoted_strings.join(" ")
+    vec![("args".to_string(), quoted_strings.join(" "))]
 }
 
 async fn rebase_mutable_descendants(
